Introduce new default enum value for text and icon rotation and pitch
This adds a new enumeration value called `auto` for
text-rotation-alignment, icon-rotation-alignment and
text-pitch-alignment. The new value is set as the default for the
three properties it was added to and the associated `doc`
description was updated to reflect that.

In addition, the relevant SDK support tables have been updated to
reflect the first expected native SDK releases that'll support the new
`auto` value. The js SDK version was not updated since the value was
not known at the time of this commit. It can be updated in the future.